+++ Sources/SwiftUIKit/Pickers/ListMultiPicker.swift
@@ -1,5 +1,5 @@
 //
-//  Picker.swift
+//  MultiPicker.swift
 //  SwiftUIKit
 //
 //  Created by Daniel Saidi on 2021-08-20.
@@ -9,47 +9,47 @@
 import SwiftUI
 
 /**
- This picker view lists an `Identifiable` item collection in
- a `List` and binds its `selection` to an external value.
+ This generic picker lists `Identifiable` items in a SwiftUI
+ `List` and binds its `selection` to an external value.
  
- The picker uses the provided `listItem` builder to build an
- item view for each item in the provided `items` collection.
- 
- If `dismissAfterPick` is `true` the picker dismisses itself
- automatically when an item is picked.
+ You can use this view instead of the native SwiftUI `Picker`
+ to get more control over the list item views. The view uses
+ the `listItem` function to build an item view for each item
+ in the provided `items` or `sections`.
  */
-public struct Picker<Item: Identifiable, ItemView: View>: View, DismissableView {
+public struct ListMultiPicker<Item: Identifiable, ItemView: View>: View, DismissableView {
     
+    /**
+     Create a picker with a single section.
+     */
     public init(
         title: String,
         items: [Item],
-        selection: Binding<Item>,
-        dismissAfterPick: Bool = true,
+        selection: Binding<[Item]>,
         listItem: @escaping ItemViewBuilder) {
         self.title = title
-        self.sections = [PickerSection(title: "", items: items)]
+        self.sections = [ListPickerSection(title: "", items: items)]
         self.selection = selection
-        self.dismissAfterPick = dismissAfterPick
         self.listItem = listItem
     }
     
+    /**
+     Create a picker with multiple sections.
+     */
     public init(
         title: String,
-        sections: [PickerSection<Item>],
-        selection: Binding<Item>,
-        dismissAfterPick: Bool = true,
+        sections: [ListPickerSection<Item>],
+        selection: Binding<[Item]>,
         listItem: @escaping ItemViewBuilder) {
         self.title = title
         self.sections = sections
         self.selection = selection
-        self.dismissAfterPick = dismissAfterPick
         self.listItem = listItem
     }
     
     private let title: String
-    private let sections: [PickerSection<Item>]
-    private let selection: Binding<Item>
-    private let dismissAfterPick: Bool
+    private let sections: [ListPickerSection<Item>]
+    private let selection: Binding<[Item]>
     private let listItem: ItemViewBuilder
     
     public typealias ItemViewBuilder = (_ item: Item, _ isSelected: Bool) -> ItemView
@@ -59,7 +59,7 @@ public struct Picker<Item: Identifiable, ItemView: View>: View, DismissableView 
     public var body: some View {
         List {
             ForEach(sections) { section in
-                Section(header: section.pickerHeader) {
+                Section(header: section.header) {
                     ForEach(section.items) {
                         listItemView(for: $0)
                     }
@@ -81,35 +81,35 @@ private extension View {
     }
 }
 
-private extension Picker {
+private extension ListMultiPicker {
     
     func listItemView(for item: Item) -> some View {
-        Button(action: { select(item) }, label: {
+        Button(action: { toggleSelection(for: item) }, label: {
             listItem(item, isSelected(item))
         })
     }
 }
 
-private extension Picker {
+private extension ListMultiPicker {
     
-    var seletedId: Item.ID {
-        selection.wrappedValue.id
+    var seletedIds: [Item.ID] {
+        selection.wrappedValue.map { $0.id }
     }
     
     func isSelected(_ item: Item) -> Bool {
-        seletedId == item.id
+        seletedIds.contains(item.id)
     }
     
-    func select(_ item: Item) {
-        selection.wrappedValue = item
-        if dismissAfterPick {
-            dismiss()
+    func toggleSelection(for item: Item) {
+        if isSelected(item) {
+            selection.wrappedValue = selection.wrappedValue.filter { $0.id != item.id }
+        } else {
+            selection.wrappedValue.append(item)
         }
     }
 }
 
-#if os(iOS) || os(tvOS)
-struct Picker_Previews: PreviewProvider {
+struct ListMultiPicker_Previews: PreviewProvider {
     
     static var previews: some View {
         Preview()
@@ -117,24 +117,24 @@ struct Picker_Previews: PreviewProvider {
     
     struct Preview: View {
         
-        @State private var selection = PreviewItem.all[0]
+        @State private var selection = [PreviewItem.all[0]]
         
-        func section(_ title: String) -> PickerSection<PreviewItem> {
-            PickerSection(title: title, items: PreviewItem.all)
+        func section(_ title: String) -> ListPickerSection<PreviewItem> {
+            ListPickerSection(
+                title: title,
+                items: PreviewItem.all)
         }
         
         var body: some View {
-            NavigationView {
-                Picker(
-                    title: "Pick something",
-                    sections: [
-                        section(""),
-                        section("Another section")
-                    ],
-                    selection: $selection) { item, isSelected in
-                        PreviewPickerItem(item: item, isSelected: isSelected)
-                    }
-            }
+            ListMultiPicker(
+                title: "Pick something",
+                sections: [
+                    section(""),
+                    section("Another section")
+                ],
+                selection: $selection) { item, isSelected in
+                    PreviewPickerItem(item: item, isSelected: isSelected)
+                }
         }
     }
     
@@ -153,7 +153,7 @@ struct Picker_Previews: PreviewProvider {
         ]
     }
 
-    struct PreviewPickerItem: View, PickerListItem {
+    struct PreviewPickerItem: View, ListPickerItem {
         
         let item: PreviewItem
         let isSelected: Bool
@@ -167,4 +167,3 @@ struct Picker_Previews: PreviewProvider {
         }
     }
 }
-#endif